Added an aspect to control the execution of scenarios with @jira tag
