+++ src/test/resources/META-INF/aop.xml
@@ -22,6 +22,7 @@
         <aspect name="com.privalia.qa.aspects.LoopIncludeTagAspect"/>
         <aspect name="com.privalia.qa.aspects.ReplacementAspect"/>
         <aspect name="com.privalia.qa.aspects.IgnoreTagAspect"/>
+        <aspect name="com.privalia.qa.aspects.JiraTagAspect"/>
     </aspects>
     <weaver
             options="-Xlint:ignore -Xset:weaveJavaPackages=true,weaveJavaxPackages=true">
